agent_framework/msagent/pyproject.toml
[project]
name = "msagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b260123",
    "python-dotenv>=1.2.1",
]
